Rename sync to sync_codec
